Clobber preexisting files w/my updated and extended ones
