+++ tests/applications/map.rkt
@@ -2,7 +2,7 @@
 
 (require "../../all.rkt")
 
-
+(record-bench 'synth/ground-context 'unstaged 'evalo-map #:description "The body of a function mapped over several examples")
 (time-test
  (run 1 (q)
    (evalo-staged
@@ -20,10 +20,9 @@
       ((a . a) (b . b)))))
    '((cons x x)))
 
-
-(time-test
- (run 1 (q)
-   (staged
+(record-bench 'synth/ground-context 'staging 'evalo-map)
+(defrel (staged-map-query q)
+  (time-staged
    (evalo-staged
     `(letrec ([map (lambda (f l)
                      (if (null? l)
@@ -37,7 +36,10 @@
     '(((1 . 1) (2 . 2) (3 . 3))
       (((1 2) . (1 2)) ((2 3) . (2 3)))
       ((a . a) (b . b))))))
+
+
+(record-bench 'synth/ground-context 'staged  'evalo-map)
+(time-test
+ (run 1 (q)
+   (staged-map-query q))
    '((cons x x)))
-
-
-
