+++ tests/interpreter/basics.rkt
@@ -140,17 +140,33 @@
          (later (== '(#f 1) q)))))
   '((#f 1)))
 
+(defrel (eval-on-append-synth-base q)
+  (staged
+   (evalo-staged
+	`(letrec ((append
+			   (lambda (xs ys)
+				 (if (null? xs) ,q
+					 (cons (car xs) (append (cdr xs) ys))))))
+	   (append '(1 2) '(3 4)))
+	'(1 2 3 4))))
+
 (test
-    (run 1 (q)
-      (staged
-       (evalo-staged
-        `(letrec ((append
-                   (lambda (xs ys)
-                     (if (null? xs) ,q
-                         (cons (car xs) (append (cdr xs) ys))))))
-           (append '(1 2) '(3 4)))
-        '(1 2 3 4))))
+  (run 1 (q)
+	(eval-on-append-synth-base q))
   '(ys))
+
+(test
+  (run 1 (q)
+	(evalo-unstaged
+	 `(letrec ((append
+				(lambda (xs ys)
+				  (if (null? xs) ,q
+					  (cons (car xs) (append (cdr xs) ys))))))
+		(append '(1 2) '(3 4)))
+	 '(1 2 3 4)))
+  '(ys))
+
+
 
 (test
     (run 1 (q)
@@ -212,6 +228,7 @@
       (append ',xs ',ys))
    zs))
 
+
 (defrel (appendo-staged xs ys zs)
   (staged
    (evalo-staged
@@ -256,10 +273,60 @@
         #:title "Handwritten/Staged/Unstaged Runtime vs List Size"
         #:out-file out-file))
 
+
 (test
-    (run* (xs ys)
-      (appendo-staged xs ys '(a b c)))
-  '((() (a b c)) ((a) (b c)) ((a b) (c)) ((a b c) ())))
+  (last-pair (run* (xs ys)
+	(appendo-staged ys xs '(A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
+							  AA AB AC AD AE AF AG AH AI AJ AK AL AM AN AO AP AQ AR AS AT AU AV AW AX AY AZ
+							  BA BB BC BD BE BF BG BH BI BJ BK BL BM BN BO BP BQ BR BS BT BU BV BW BX BY BZ
+							  CA CB CC CD CE CF CG CH CI CJ CK CL CM CN CO CP CQ CR CS CT CU CV CW CX CY CZ
+							  DA DB DC DD DE DF DG DH DI DJ DK DL DM DN DO DP DQ DR DS DT DU DV DW DX DY DZ
+							  EA EB EC ED EE EF EG EH EI EJ EK EL EM EN EO EP EQ ER ES ET EU EV EW EX EY EZ
+							  FA FB FC FD FE FF FG FH FI FJ FK FL FM FN FO FP FQ FR FS FT FU FV FW FX FY FZ
+							  GA GB GC GD GE GF GG GH GI GJ GK GL GM GN GO GP GQ GR GS GT GU GV GW GX GY GZ
+							  HA HB HC HD HE HF HG HH HI HJ HK HL HM HN HO HP HQ HR HS HT HU HV HW HX HY HZ
+							  IA IB IC ID IE IF IG IH II IJ IK IL IM IN IO IP IQ IR IS IT IU IV IW IX IY IZ
+							  JA JB JC JD JE JF JG JH JI JJ JK JL JM))))
+'((()
+   (A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
+	AA AB AC AD AE AF AG AH AI AJ AK AL AM AN AO AP AQ AR AS AT AU AV AW AX AY AZ
+	BA BB BC BD BE BF BG BH BI BJ BK BL BM BN BO BP BQ BR BS BT BU BV BW BX BY BZ
+	CA CB CC CD CE CF CG CH CI CJ CK CL CM CN CO CP CQ CR CS CT CU CV CW CX CY CZ
+	DA DB DC DD DE DF DG DH DI DJ DK DL DM DN DO DP DQ DR DS DT DU DV DW DX DY DZ
+	EA EB EC ED EE EF EG EH EI EJ EK EL EM EN EO EP EQ ER ES ET EU EV EW EX EY EZ
+	FA FB FC FD FE FF FG FH FI FJ FK FL FM FN FO FP FQ FR FS FT FU FV FW FX FY FZ
+	GA GB GC GD GE GF GG GH GI GJ GK GL GM GN GO GP GQ GR GS GT GU GV GW GX GY GZ
+	HA HB HC HD HE HF HG HH HI HJ HK HL HM HN HO HP HQ HR HS HT HU HV HW HX HY HZ
+	IA IB IC ID IE IF IG IH II IJ IK IL IM IN IO IP IQ IR IS IT IU IV IW IX IY IZ
+	JA JB JC JD JE JF JG JH JI JJ JK JL JM))))
+
+
+(test
+  (last-pair (run* (xs ys)
+	(appendo-unstaged ys xs '(A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
+							  AA AB AC AD AE AF AG AH AI AJ AK AL AM AN AO AP AQ AR AS AT AU AV AW AX AY AZ
+							  BA BB BC BD BE BF BG BH BI BJ BK BL BM BN BO BP BQ BR BS BT BU BV BW BX BY BZ
+							  CA CB CC CD CE CF CG CH CI CJ CK CL CM CN CO CP CQ CR CS CT CU CV CW CX CY CZ
+							  DA DB DC DD DE DF DG DH DI DJ DK DL DM DN DO DP DQ DR DS DT DU DV DW DX DY DZ
+							  EA EB EC ED EE EF EG EH EI EJ EK EL EM EN EO EP EQ ER ES ET EU EV EW EX EY EZ
+							  FA FB FC FD FE FF FG FH FI FJ FK FL FM FN FO FP FQ FR FS FT FU FV FW FX FY FZ
+							  GA GB GC GD GE GF GG GH GI GJ GK GL GM GN GO GP GQ GR GS GT GU GV GW GX GY GZ
+							  HA HB HC HD HE HF HG HH HI HJ HK HL HM HN HO HP HQ HR HS HT HU HV HW HX HY HZ
+							  IA IB IC ID IE IF IG IH II IJ IK IL IM IN IO IP IQ IR IS IT IU IV IW IX IY IZ
+							  JA JB JC JD JE JF JG JH JI JJ JK JL JM))))
+'((()
+   (A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
+	AA AB AC AD AE AF AG AH AI AJ AK AL AM AN AO AP AQ AR AS AT AU AV AW AX AY AZ
+	BA BB BC BD BE BF BG BH BI BJ BK BL BM BN BO BP BQ BR BS BT BU BV BW BX BY BZ
+	CA CB CC CD CE CF CG CH CI CJ CK CL CM CN CO CP CQ CR CS CT CU CV CW CX CY CZ
+	DA DB DC DD DE DF DG DH DI DJ DK DL DM DN DO DP DQ DR DS DT DU DV DW DX DY DZ
+	EA EB EC ED EE EF EG EH EI EJ EK EL EM EN EO EP EQ ER ES ET EU EV EW EX EY EZ
+	FA FB FC FD FE FF FG FH FI FJ FK FL FM FN FO FP FQ FR FS FT FU FV FW FX FY FZ
+	GA GB GC GD GE GF GG GH GI GJ GK GL GM GN GO GP GQ GR GS GT GU GV GW GX GY GZ
+	HA HB HC HD HE HF HG HH HI HJ HK HL HM HN HO HP HQ HR HS HT HU HV HW HX HY HZ
+	IA IB IC ID IE IF IG IH II IJ IK IL IM IN IO IP IQ IR IS IT IU IV IW IX IY IZ
+	JA JB JC JD JE JF JG JH JI JJ JK JL JM))))
+
 
 (test
     (run 1 (q)
